Make a materialized view for the specific query we want to display
The current algorithm is:
- gist index on journeypattern_boundingbox to get journeypattern_id
- for each journeypattern_id, index scan jptiminglink to get timing info

But but there are thousands of trivial journeypattern_id variants,
escpecially in london, which means doing the jptiminglink index
scan a *lot* of times, and this is slow.

Instead, we could sum() the frequencies of all the bus stop pairs,
and store that in a matview. Means we query using one gist index,
at the cost of creating an extra 1GB table.