Change copyright from AGPLv3 to GPLv3
This is a loosening of the copyright restrictions on the project.

The AGPL probably also applies to any transformed data, and the
results of any API queries. This makes the API harder to use, and
may have strange edge-cases if the AGPL and OGL licenses interact.

Using the GPL instead makes the situation easier: the software for
transforming, serving and displaying the data is GPL licensed; and
the resulting data is OGL licensed.