Use a model for bus acceleration and deceleration
Old method assumed that the averge speed of the bus is the top speed
the bus would accelerate to.

ie: we assume that the majority of the time is not accelerating and
decelerating. Then we find places where reducing the time spent
accelerating and decelerating would help!

This patch tries to work out the speed of the bus travels between
bus stops, adjusting for the time spent accelerating and decelerating.